Big Change
The GUI is created =
Current Used : Click the red start here button, the timer begins, it can
be paused and resumed with the button on the top right, the mainText's
text should be replaced by the generated question, the answers's text
should be replaced by the generated answers. The user can choose
multiple answers, we'll probably change this tomorrow for ease of use
and programming.  Once the options are selected the user can press
submit to get a new set of questions and answers to submit again. When
the timer is done we'll create an activity to display statistics.